group sections to match figma and fix steps
