--- apps/cxc/components/application/steps/preferences.tsx
@@ -1,103 +1,0 @@
-"use client";
-
-import {
-  Card,
-  CardHeader,
-  CardTitle,
-  CardContent,
-  Form,
-  FormField,
-} from "@uwdsc/ui";
-
-import { UseFormReturn } from "react-hook-form";
-import { AppFormValues } from "@/lib/schemas/application";
-import {
-  renderTextFieldWithLabel as renderTextField,
-  renderSelectField,
-} from "@/components/FormHelpers";
-
-import { Shirt } from "lucide-react"; 
-
-import { useEffect } from "react";
-
-interface PreferencesProps {
-  readonly form: UseFormReturn<AppFormValues>;
-}
-
-// Preferences options
-const DIETARY_OPTIONS = [
-  "None",
-  "Vegetarian",
-  "Vegan",
-  "Gluten-Free",
-  "Halal",
-  "Kosher",
-  "Other",
-];
-
-const TSHIRT_OPTIONS = ["XS", "S", "M", "L", "XL", "XXL"];
-
-export function Preferences({ form }: PreferencesProps) {
-  const dietaryRestriction = form.watch("dietary_restrictions");
-  const tShirtSize = form.watch("tshirt_size");
-
-  useEffect(() => {
-    if (dietaryRestriction !== "Other") {
-      form.setValue("dietary_restrictions_other", "");
-    }
-  }, [dietaryRestriction, form]);
-
-  return (
-    <div className="space-y-6">
-      <Form {...form}>
-        <Card className="bg-zinc-900/50">
-          <CardHeader>
-            <CardTitle className="flex items-center text-lg font-semibold">
-              About You
-            </CardTitle>
-          </CardHeader>
-          <CardContent className="space-y-6">
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-                {/* Dietary Restrictions */}
-                <div className="flex flex-col space-y-4">
-                    <FormField
-                    control={form.control}
-                    name="dietary_restrictions"
-                    render={renderSelectField(
-                        "Dietary Restrictions",
-                        DIETARY_OPTIONS,
-                        {label: "Select any dietary restriction", required: true},
-                    )}
-                    />
-                    
-                    {dietaryRestriction === "Other" && (
-                    <FormField
-                        control={form.control}
-                        name="dietary_restrictions_other"
-                        render={renderTextField(
-                        "Other Dietary Restriction",
-                        "Enter custom restriction"
-                        )}
-                    />
-                    )}  
-                </div>
-            
-                {/* T-Shirt Size */}
-                <div>
-                    <FormField
-                    control={form.control}
-                    name="tshirt_size"
-                    render={renderSelectField(
-                        "T-Shirt Size",
-                        TSHIRT_OPTIONS,
-                        {label: "Select your T-shirt size...", required:true},
-                    )}
-                    />
-                </div>
-            </div> 
-          </CardContent>
-        </Card>
-      </Form>
-    </div>
-  );
-}
