+++ apps/cxc/lib/utils/application.ts
@@ -15,20 +15,41 @@ export const isStepValid = (
 
   switch (currentStep) {
     case 1: // Personal Details
+      const dietary_restrictions = form.watch("dietary_restrictions");
+      const dietary_restrictions_other = form.watch(
+        "dietary_restrictions_other"
+      );
+      const tshirt_size = form.watch("tshirt_size");
       const gender = form.watch("gender");
       const ethnicity = form.watch("ethnicity");
-      return (
+
+      const validAboutYou =
+        !errors.dietary_restrictions &&
+        dietary_restrictions !== undefined &&
+        !errors.tshirt_size &&
+        tshirt_size !== undefined &&
+        !errors.dietary_restrictions_other &&
+        (dietary_restrictions !== "Other" ||
+          (dietary_restrictions_other?.trim().length ?? 0) > 0);
+
+      const validContactInfo =
         !errors.email &&
         !errors.phone &&
         !errors.discord &&
         gender !== undefined &&
-        ethnicity !== undefined
-      );
-    case 2: // Education
+        ethnicity !== undefined;
+
+      return validAboutYou && validContactInfo;
+    case 2: // Experience
       const universityName = form.watch("university_name");
       const program = form.watch("program");
       const universityNameOther = form.watch("university_name_other");
       const programOther = form.watch("program_other");
+
+      const prior_hackathon_experience = form.watch(
+        "prior_hackathon_experience"
+      );
+      const hackathons_attended = form.watch("hackathons_attended");
 
       const isUniversityValid =
         !errors.university_name &&
@@ -41,29 +62,7 @@ export const isStepValid = (
       const isYearValid =
         !errors.year_of_study && !!form.watch("year_of_study");
 
-      return isUniversityValid && isProgramValid && isYearValid;
-    case 3:
-      const dietary_restrictions = form.watch("dietary_restrictions");
-      const dietary_restrictions_other = form.watch(
-        "dietary_restrictions_other"
-      );
-      const tshirt_size = form.watch("tshirt_size");
-      return (
-        !errors.dietary_restrictions &&
-        dietary_restrictions !== undefined &&
-        !errors.tshirt_size &&
-        tshirt_size !== undefined &&
-        !errors.dietary_restrictions_other &&
-        (dietary_restrictions !== "Other" ||
-          (dietary_restrictions_other?.trim().length ?? 0) > 0)
-      );
-    case 4:
-      const prior_hackathon_experience = form.watch(
-        "prior_hackathon_experience"
-      );
-      console.log(prior_hackathon_experience);
-      const hackathons_attended = form.watch("hackathons_attended");
-      return (
+      const validHackExp =
         !errors.prior_hackathon_experience &&
         !errors.resume &&
         !errors.github &&
@@ -71,10 +70,11 @@ export const isStepValid = (
         !errors.other_link &&
         !errors.hackathons_attended &&
         prior_hackathon_experience.length > 0 &&
-        hackathons_attended !== undefined
-        // TODO: ^ change to number input instead of dropdown later
-      );
-    case 5: // CxC App
+        hackathons_attended !== undefined;
+      // TODO: ^ change to number input instead of dropdown later
+
+      return isUniversityValid && isProgramValid && isYearValid && validHackExp;
+    case 3: // CxC App
       return !errors.cxc_gain && !errors.silly_q;
     default:
       return true;
